zaimplementowane metody GetMinimalCost, GetCosts z interfejsu

+++ Model.Bee/TestData.cs
@@ -21,7 +21,7 @@ namespace Model.Bee
             this.n = problemSize;
         }
 
-        public double Cost(int l1, int l2, int f1, int f2)
+        public int Cost(int l1, int l2, int f1, int f2)
         {
             if(l1 < n && l2 < n && f1 < n && f2 < n)
                 return A[f1,f2]*B[l1,l2];
